Extract example token request creation code to a separate class for clarity
